# api/storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "CraftedBy-api Documentation",
        "description": "CraftedBy api",
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://localhost:8000",
            "description": "local server"
        },
        {
            "url": "http://staging.example.com",
            "description": "staging server"
        },
        {
            "url": "http://example.com",
            "description": "production server"
        }
    ],
    "paths": {
        "/api/products": {
            "get": {
                "tags": [
                    "Products"
                ],
                "summary": "List all products",
                "operationId": "d97d1c3727cbac49708b85f7d11f7d58",
                "responses": {
                    "201": {
                        "description": "Status Ok"
                    },
                    "500": {
                        "description": "Server Error"
                    }
                },
                "security": []
            }
        },
        "/api/products/create": {
            "post": {
                "tags": [
                    "Product"
                ],
                "summary": "Create a new Product",
                "operationId": "1b55d35de914dc6a5755c548182f4fce",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/x-www-form-urlencoded": {
                            "schema": {
                                "required": [
                                    "name",
                                    "price",
                                    "description",
                                    "material",
                                    "category",
                                    "artist"
                                ],
                                "properties": {
                                    "name": {
                                        "description": "Product name",
                                        "type": "string"
                                    },
                                    "description": {
                                        "description": "Product description",
                                        "type": "text"
                                    },
                                    "price": {
                                        "description": "Product price",
                                        "type": "decimal"
                                    },
                                    "material": {
                                        "description": "Product Material",
                                        "type": "string"
                                    },
                                    "category": {
                                        "description": "Product Category",
                                        "type": "string"
                                    },
                                    "artist": {
                                        "description": "Artist who create the product",
                                        "type": "string"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "401": {
                        "description": "Unauthorized"
                    },
                    "201": {
                        "description": "Register Successfully"
                    },
                    "422": {
                        "description": "Unprocessable entity"
                    },
                    "500": {
                        "description": "Server Error"
                    }
                },
                "security": {
                    "bearerAuth": {}
                }
            }
        }
    },
    "components": {
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "name": "Authorization",
                "in": "header",
                "scheme": "bearer"
            }
        }
    }
}